changed shape_dist_traveled from int to float for shapes. I wonder if it's in a different unit than the others?

+++ infrastructure/db/schema.sql
@@ -32,7 +32,7 @@ create table cta_tracker.shapes (
   shape_pt_lat float,
   shape_pt_lon float,
   shape_pt_sequence int,
-  shape_dist_traveled int
+  shape_dist_traveled float
 );
 create table cta_tracker.stop_times (
   trip_id text,
